+++ mojo/__init__.py
@@ -1,3 +1,3 @@
-__version__ = "1.2.15"
+__version__ = "1.2.16"
 
 from mojo.helpers.response import JsonResponse